File structure reorganized (using:index.js|index.test.js|style.scss)
