+++ src/components/atoms/Item/style.scss
@@ -1,0 +1,3 @@
+.item {
+  background-color: green;
+}
