--- src/components/atoms/Label/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import styles from './styles.scss';
-
-const Label = () => <div className={styles.app}>{'Hello world :)'}</div>;
-
-export default Label;
