--- src/components/atoms/Label/styles.scss
@@ -1,3 +1,0 @@
-.app {
-  background-color: green;
-}
