+++ src/components/ecosystems/LoginScreen/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { pushRoute, getPost, clearPost } from '../redux/modules/app';
-import { login } from '../redux/modules/user';
+import { pushRoute, getPost, clearPost } from './../../../redux/modules/App/index';
+import { login } from './../../../redux/modules/User/index';
 
 class LoginScreen extends React.Component {
   static propTypes = {
