+++ src/components/molecules/List/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Item from './Item';
+import Item from './../../atoms/Item/index';
 
 const List = props => (
   <ul>
