+++ src/redux/modules/App/index.js
@@ -1,6 +1,6 @@
 import { push } from 'react-router-redux';
 
-import * as types from '../types';
+import * as types from './../../types';
 
 // Action Creators
 export function pushRoute(route) {
